% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{languages}
\alias{languages}
\title{Language Models.}
\format{
A data frame with 60 observations and 2 variables:
\code{short}, and \code{repo}.
}
\description{
Known language models for lemmatizzation and POS tagging through udpipe models.
}
