% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reinert.R
\name{merge_small_segments}
\alias{merge_small_segments}
\title{merge small segments with the previous ones}
\usage{
merge_small_segments(idTable, min_length = 5)
}
\arguments{
\item{idTable}{is a data frame obtained by split_segments}

\item{min_length}{min number of forms by segment. Default value is segment_size = 5}
}
\value{
ordered list of document indices
}
\description{
merge small segments with the previous ones
}
\details{
Internal function, not to be used directly
}
