% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reinert.R
\name{split_segments}
\alias{split_segments}
\title{split documents into segments}
\usage{
split_segments(doc_id, segment_size)
}
\arguments{
\item{doc_id}{is the  vector of doc_id labels extracted from x}

\item{segment_size}{number of forms by document. Default value is segment_size = 40}
}
\value{
ordered list of document indices
}
\description{
split documents into segments
}
\details{
Internal function, not to be used directly
}
