% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txt_recode_fast.R
\name{txt_recode_cpp}
\alias{txt_recode_cpp}
\title{Internal C++ text recoding function}
\arguments{
\item{x}{Character vector to recode}

\item{from}{Character vector of values to match}

\item{to}{Character vector of replacement values}

\item{na_rm}{Logical, whether to set unmatched values to NA}
}
\value{
Character vector with recoded values
}
\description{
Internal C++ implementation of text recoding using hash tables.
Users should call \code{\link{txt_recode_fast}} instead.
}
\keyword{internal}
